% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{dgbetapr}
\alias{dgbetapr}
\title{PDF of Generalized Beta Prime distribution}
\usage{
dgbetapr(x, alpha, beta, c, d)
}
\arguments{
\item{x}{vector of quantiles}

\item{alpha}{shape parameter}

\item{beta}{shape parameter}

\item{c}{power parameter}

\item{d}{scale parameter}
}
\value{
value of pdf at x
}
\description{
This function computes the value of the
probability density function of a generalized
beta prime (GBP) distribution
}
