% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{pgbetapr}
\alias{pgbetapr}
\title{CDF of Generalized Beta Prime distribution}
\usage{
pgbetapr(x, alpha, beta, c, d)
}
\arguments{
\item{x}{vector of quantiles}

\item{alpha}{shape parameter}

\item{beta}{shape parameter}

\item{c}{power parameter}

\item{d}{scale parameter}
}
\value{
value of cdf at x
}
\description{
This function computes the value of the
distribution function of a generalized
beta prime (GBP) distribution
}
