% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{qgbetapr}
\alias{qgbetapr}
\title{Quantile Function of Generalized Beta Prime distribution}
\usage{
qgbetapr(p, alpha, beta, c, d)
}
\arguments{
\item{p}{vector of probabilities}

\item{alpha}{shape parameter}

\item{beta}{shape parameter}

\item{c}{power parameter}

\item{d}{scale parameter}
}
\value{
quantile corresponding to p
}
\description{
This function computes the quantile of a generalized
beta prime (GBP) distribution
}
