% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{qw}
\alias{qw}
\title{Quantile Function of W}
\usage{
qw(
  p,
  a = 1,
  b = 1,
  b0 = 0,
  link = c("Gaussian", "Logistic", "Poisson", "Poisson_offsets", "ZIP", "NegBinom",
    "Arbitrary"),
  theta = 1,
  mu,
  sigma2,
  disp = FALSE
)
}
\arguments{
\item{p}{vector of probabilities}

\item{a}{value of a for prior R2~Beta(a,b)}

\item{b}{value of b for prior R2~Beta(a,b)}

\item{b0}{value of beta0}

\item{link}{link function used to compute R2. Default options are "Gaussian", "Logistic",
"Poisson", "Poisson_offsets", "ZIP" and "NegBinom". Also can have user-inputted link
function using "Arbitrary".}

\item{theta}{dispersion parameter (if necessary)}

\item{mu}{if link="Arbitrary", then this is the mean function of the linear predictor.
May depend on theta}

\item{sigma2}{if link="Arbitrary", then this is the variance function of the linear predictor.
May depend on theta}

\item{disp}{logical. TRUE if arbitrary link function depends on a dispersion parameter}
}
\value{
quantile of W at p
}
\description{
This function computes quantile of W induced by a
Beta(a,b) prior on R2 for any link function
}
