% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{rgbetapr}
\alias{rgbetapr}
\title{Random Number Generator of Generalized Beta Prime distribution}
\usage{
rgbetapr(n, alpha, beta, c, d)
}
\arguments{
\item{n}{number of observations}

\item{alpha}{shape parameter}

\item{beta}{shape parameter}

\item{c}{power parameter}

\item{d}{scale parameter}
}
\value{
quantile corresponding to p
}
\description{
This function generates a random number from a generalized
beta prime (GBP) distribution
}
