% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{W_to_R2}
\alias{W_to_R2}
\title{R2 for a given W}
\usage{
W_to_R2(
  W,
  b0 = 0,
  link = c("Gaussian", "Logistic", "Poisson", "Poisson_offsets", "ZIP", "NegBinom",
    "Arbitrary"),
  theta = 1,
  mu,
  sigma2,
  disp = FALSE
)
}
\arguments{
\item{W}{value of W}

\item{b0}{value of beta0}

\item{link}{link function used to compute R2. Default options are "Gaussian", "Logistic",
"Poisson", "Poisson_offsets", "ZIP" and "NegBinom". Also can have user-inputted link
function using "Arbitrary".}

\item{theta}{dispersion parameter (if necessary)}

\item{mu}{if link="Arbitrary", then this is the mean function of the linear predictor.
May depend on theta}

\item{sigma2}{if link="Arbitrary", then this is the variance function of the linear predictor.
May depend on theta}

\item{disp}{logical. TRUE if arbitrary link function depends on a dispersion parameter}
}
\value{
value of R2
}
\description{
This function computes the value of R2 for
a given W, beta0 and link function
}
