% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{WGBP}
\alias{WGBP}
\title{GBP Approximation (Logistic)}
\usage{
WGBP(
  a = 1,
  b = 1,
  lambda = 0.25,
  b0 = 0,
  link = c("Logistic", "Poisson", "Poisson_offsets", "ZIP", "NegBinom", "Arbitrary"),
  theta = 1,
  mu,
  sigma2,
  disp = FALSE
)
}
\arguments{
\item{a}{value of a for prior R2~Beta(a,b)}

\item{b}{value of b for prior R2~Beta(a,b)}

\item{lambda}{tuning parameter}

\item{b0}{value of beta0}

\item{link}{link function used to compute R2. Default options are "Logistic",
"Poisson", "Poisson_offsets", "ZIP" and "NegBinom". Also can have user-inputted link
function using "Arbitrary".}

\item{theta}{dispersion parameter (if necessary)}

\item{mu}{if link="Arbitrary", then this is the mean function of the linear predictor.
May depend on theta}

\item{sigma2}{if link="Arbitrary", then this is the variance function of the linear predictor.
May depend on theta}

\item{disp}{logical. TRUE if arbitrary link function depends on a dispersion parameter}
}
\value{
parameters (alpha,beta,c,d) for GBP
}
\description{
This function finds the closest generalized beta prime (GBP)
distribution to the true pdf of W as measured by the
Pearson chi-squared divergence
}
